requirements.txt
django==2.1
mysqlclient==1.3.12
django-redis==4.11.0
djangorestframework==3.11.0
gunicorn==20.0.4
openpyxl==3.0.3
Pillow==7.1.1
patool==1.12
django-cors-headers==3.3.0







